+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## 0.19.3
+
+### Added
+
+- Schemas : Implement Annotated Schemas for type constraints
+
+### Changed
+
+- Workflow : Block MultiObject configurations type does not allow base to be instantiated anymore
+
 ## 0.19.2
 
 ### Changed